groq_rag/requirements.txt
langchain
langchain_core
langchain_community
langchain_openai
llama-parse
fastembed
chromadb
python-dotenv
langchain-groq
chainlit
fastembed
unstructured
langserve
streamlit
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
cassio
langchainhub
markdown